Create stats in HistoryCallback
